utils: refactor platform/stop_signal.h into utils
This was the one and only file in platform/: it
doesn't deserve a whole top level directory.
