+++ src/v/raft/kvelldb/kvserver.cc
@@ -12,7 +12,6 @@
 #include "model/metadata.h"
 #include "net/server.h"
 #include "net/unresolved_address.h"
-#include "platform/stop_signal.h"
 #include "raft/consensus.h"
 #include "raft/consensus_client_protocol.h"
 #include "raft/heartbeat_manager.h"
@@ -30,6 +29,7 @@
 #include "storage/logger.h"
 #include "syschecks/syschecks.h"
 #include "utils/hdr_hist.h"
+#include "utils/stop_signal.h"
 #include "vlog.h"
 
 #include <seastar/core/app-template.hh>
