+++ src/v/utils/stop_signal.h
